+++ composeApp/src/desktopMain/kotlin/ai/mcpdirect/studio/app/mcp/MCPServerNotificationHandlerImplement.kt
@@ -3,7 +3,7 @@ package ai.mcpdirect.studio.app.mcp
 import ai.mcpdirect.studio.dao.entity.MCPServer
 import ai.mcpdirect.studio.handler.MCPServerNotificationHandler
 
-class MCPServerNotificationHandlerImplement(private val viewModel: MCPServerIntegrationViewModel) : MCPServerNotificationHandler {
+class MCPServerNotificationHandlerImplement(private val viewModel: ConnectMCPViewModel) : MCPServerNotificationHandler {
 
     override fun onMCPServersNotification(servers: List<MCPServer>) {
         viewModel.updateServer(servers)